+++ packages/york-web/src/components/complex/Header/DesktopHeader/Profile/index.js
@@ -7,12 +7,12 @@ import { Text, View } from 'york-web/components/primitive'
 
 import { menuItemShape, componentsShape } from '../../utils'
 
+import locales from '../../locales'
 import LoginIcon from '../../assets/login.svg'
 import ProfileIcon from '../../assets/profile.svg'
 import ProfilePlusIcon from '../../assets/profilePlus.svg'
 
 import Dropdown from './Dropdown'
-import locales from './locales'
 
 const StyledLogin = styled(View)`
   user-select: none;
