--- packages/york-web/src/components/complex/Header/MobileHeader/MobileBurgerHeader/locales.js
@@ -1,4 +1,0 @@
-export default {
-  login: 'Войти',
-  profile: 'Профиль',
-}
